cloudfunction/get_alt_text/requirements.txt
functions-framework==3.*
flask
vertexai
urllib3
google-cloud-storage